Refactor: Rename app package to ghapp
The app package contains code for a GitHub App

* Per https://github.com/jlewi/hydros/blob/main/docs/design_declarative_cd.md#github-app
  and #5 I don't think we will use most of this

* We'd like to have an App class to encapsulate all of hydros.
* So we rename the app package so we can keep the code until we clean it up.
