+++ images.yaml
@@ -19,8 +19,8 @@ spec:
     # Pull in kustomize from its docker image
     - uri: docker://registry.k8s.io/kustomize/kustomize:v5.0.0
       mappings:
-        - src: app/kustomize
-          strip: app
+        - src: ghapp/kustomize
+          strip: ghapp
   builder:
     gcb:
       project: foyle-public
