--- pkg/app/doc.go
@@ -1,2 +1,0 @@
-// Package app defines the actual HTTP application that handles webhooks for the GitHub App.
-package app
