+++ pkg/ghapp/doc.go
@@ -1,0 +1,2 @@
+// Package ghapp defines the actual HTTP application that handles webhooks for the GitHub App.
+package ghapp
