+++ pkg/ghapp/fetcher.go
@@ -1,4 +1,4 @@
-package app
+package ghapp
 
 import (
 	"context"
