+++ pkg/hydros/constants.go
@@ -1,6 +1,6 @@
 package hydros
 
 const (
-	// HydrosGitHubAppID is the app id for Hydros.
+	// HydrosGitHubAppID is the ghapp id for Hydros.
 	HydrosGitHubAppID = 266158
 )
